Massive refactoring into better directory structure

+++ install.sh
@@ -2,6 +2,13 @@
 
 RICE=$(pwd)
 
+dry=true
+if [ "$1" = "--install" ]; then
+  dry=false
+else
+  echo "Dry run. Run again with --install to actually install things"
+fi
+
 ensureDir() {
   if [ -e "$1" ]; then
     if ! [ -d "$1" ]; then
@@ -19,16 +26,20 @@ installAs() {
   destDir="$(dirname "$2")"
   destName="$(basename "$2")"
   printf "  Installing %-22s to %-40s as %s\n" "$srcFilename" "$destDir" "$destName"
-  ensureDir "$destDir"
-  ln -sfT "$1" "$destDir/$destName"
+  if ! $dry; then
+    ensureDir "$destDir"
+    ln -sfT "$1" "$destDir/$destName"
+  fi
 }
 
 install() {
   filename="$(basename "$1")"
   printf "  Installing %-22s to %s\n" "$filename" "$2"
-  ensureDir "$2"
-  rm -rf "$2/$filename"
-  ln -sfT "$1" "$2/$filename"
+  if ! $dry; then
+    ensureDir "$2"
+    rm -rf "$2/$filename"
+    ln -sfT "$1" "$2/$filename"
+  fi
 }
 
 # Make sure we have plug.kak, etc
@@ -46,30 +57,23 @@ mkdir -p ~/.local/share/qutebrowser-containers/
 sed "s/CONTAINER/container/g" $RICE/qutebrowser-template.desktop \
   > ~/.local/share/applications/qutebrowser-container.desktop
 
-echo "Installing items..."
-install   $RICE/clang-format          ~/
-install   $RICE/fish                  ~/.config/
-install   $RICE/kak                   ~/.config/
-install   $RICE/.profile              ~/
-install   $RICE/bemenu.sh             ~/bin
-install   $RICE/awesome               ~/.config
-install   $RICE/sway                  ~/.config
-install   $RICE/launcher.sh           ~/bin
-install   $RICE/luakit                ~/.config
-install   $RICE/kitty/                ~/.config 
-install   $RICE/waybar                ~/.config
-install   $RICE/wallpaper.sh          ~/bin
-install   $RICE/wldash                ~/.config
-installAs $RICE/vimrc                 ~/.vimrc
+echo "Installing bin stuff..."
+for file in $RICE/bin/* ; do
+  install "$file" ~/bin/
+done
+
+echo "Installing homedir dots..."
+for file in $RICE/home-dots/* ; do
+  name="$(basename "$file")"
+  installAs "$file" "$HOME/"."$name"
+done
+
+echo "Installing .config stuff..."
+for file in $RICE/config/* ; do
+  install "$file" ~/.config
+done
+
 installAs $RICE/quteconfig.py         ~/.config/qutebrowser/config.py
-install   $RICE/qutebrowser-container ~/bin/
-install   $RICE/mpd.conf              ~/.config/mpd/
-install   $RICE/mvi                   ~/bin
-install   $RICE/notify-at             ~/bin
-installAs $RICE/mako                  ~/.config/mako/config
-install   $RICE/ncmpcpp               ~/.ncmpcpp/
-install   $RICE/kickbar               ~/bin/
-install   $RICE/zathura               ~/.config/
 
 
 echo "/etc/issue requires manual installation"
